pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "email-validator>=2.2.0",
    "fastapi>=0.123.4",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "motor>=3.6.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pytesseract>=0.3.13",
    "pyjwt>=2.9.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "faiss-cpu>=1.7.4",
    "uvicorn>=0.38.0",
]
